feat: arreglos en carrito de compras

+++ project/producto/models.py
@@ -4,7 +4,7 @@ from django.utils import timezone
 class Producto(models.Model):
     nombre = models.CharField(max_length=255)
     descripcion = models.TextField()
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.FloatField()
     cantidad = models.FloatField()
     imagen = models.ImageField(upload_to='productos/', null=True, blank=True)
     fecha_actualizacion = models.DateField(null=True, blank=True, editable=False, default=timezone.now)
